backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.4
asyncpg==0.30.0
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.20
bcrypt==4.3.0
chromadb==1.0.17
langchain==0.3.27
langchain-google-genai
google-generativeai
pandas==2.3.1
scikit-learn==1.7.1
numpy==2.3.2
python-dotenv==1.0.1
pytest==8.4.1
httpx==0.28.1
redis==5.0.1
langchain-community
langchain-core
langchain-text-splitters
langchain-chroma
langchain-experimental
unstructured
openpyxl
tiktoken
tenacity
